Move create_header to text_utils file.
