+++ src/text_utils.rs
@@ -1,4 +1,6 @@
 pub mod text_utils {
+    use crate::colors::{Color, format_color};
+
     pub fn hello_mod() {
         println!("Matchcollor has been called");
     }
@@ -10,6 +12,17 @@ pub mod text_utils {
             return_value.push_str("=");
         }
         return return_value;
+    }
+
+    pub fn create_header(text: &str, color: Color) {
+        let text_length: usize = text.len() + 8;
+        let line = format_color(&write_line(text_length), &color);
+    
+        let column_line = format_color("||", &color);
+    
+        println!("{}", line);
+        println!("{column_line}  {}  {column_line}", format_color(text, &color));
+        println!("{line}");
     }
 }
